+++ doc/CHANGELOG.md
@@ -29,7 +29,8 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
-* `#2566` [Timelines] Searching when selecting columns for a timeline configuration does not work 
+* `#2566` [Timelines] Searching when selecting columns for a timeline configuration does not work
+* `#2631` [Timelines] Work package cannot be created out of timeline.
 * `#2686` [Work package tracking] Work package summary not displayed correctly
 
 ## 3.0.0pre27