+++ doc/CHANGELOG.md
@@ -29,9 +29,10 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
-* `#2566` [Timelines] Searching when selecting columns for a timeline configuration does not work
-* `#2631` [Timelines] Work package cannot be created out of timeline.
-* `#2686` [Work package tracking] Work package summary not displayed correctly
+* `#2566` Fix: [Timelines] Searching when selecting columns for a timeline configuration does not work
+* `#2631` Fix: [Timelines] Work package cannot be created out of timeline.
+* `#2686` Fix: [Work package tracking] Work package summary not displayed correctly
+* `#2687` Fix: [Work Package Tracking] No error for parallel editing
 
 ## 3.0.0pre27